Changed Login component to an emerging modal
