+++ package.json
@@ -19,6 +19,7 @@
     "react-dom": "^16.12.0",
     "react-leaflet": "^2.6.1",
     "react-scripts": "3.3.0",
+    "reactjs-popup": "^1.5.0",
     "shx": "0.3.2",
     "solid": "^0.2.1",
     "solid-auth-client": "^2.4.1",
