+++ src/components/NavBar/Login/index.js
@@ -1,0 +1,4 @@
+//import LoginComponent from "./login.component";
+import LoginModal from "./login-modal.component";
+
+export default LoginModal;
