--- src/containers/Login/index.js
@@ -1,3 +1,0 @@
-import LoginComponent from "./login.component";
-
-export default LoginComponent;
